+++ fs/ecryptfs/miscdev.c
@@ -423,8 +423,8 @@ ecryptfs_miscdev_write(struct file *file, const char __user *buf,
 		printk(KERN_WARNING "%s: Acceptable packet size range is "
 		       "[%d-%lu], but amount of data written is [%zu].",
 		       __func__, (1 + 4 + 1),
-		       (1 + 4 + 2 + sizeof(struct ecryptfs_message) + 4
-			+ ECRYPTFS_MAX_ENCRYPTED_KEY_BYTES), count);
+		       (unsigned long int)(1 + 4 + 2 + sizeof(struct ecryptfs_message)
+				   + 4 + ECRYPTFS_MAX_ENCRYPTED_KEY_BYTES), count);
 		return -EINVAL;
 	}
